refactor: make AudioPlayButton a common widget
